moved component out but is ugly TODO fix
