+++ src/app/learnmoves/data/learn/page.tsx
@@ -2,14 +2,13 @@
 import {
   useState,
   useEffect,
-  SetStateAction,
-  Dispatch,
   Suspense,
   MouseEventHandler,
 } from 'react'
 import { useLocalStorage } from '@/app/_utils/lib'
 import { MovementGroup, lsUserLearning } from '@/app/_utils/localStorageTypes'
 import { Position, Transition } from '@/app/_utils/localStorageTypes'
+import RenderMovementGroup from './MovementGroup'
 import {
   getLocalStorageGlobal,
   setLocalStorageGlobal,
@@ -20,7 +19,7 @@ import LoadingFallback from '@/app/_components/LoadingFallback'
 import { useRouter } from 'next/navigation'
 import { RenderEditButton, RenderRedDeleteButton } from '../../_components/Svgs'
 import DefaultStyledInput from '@/app/_components/DefaultStyledInput'
-import { useForm, SubmitHandler } from 'react-hook-form'
+import { useForm, SubmitHandler, Form } from 'react-hook-form'
 import {
   makeDefaultMovementGroupArr,
   makeDefaultPosition,
@@ -30,6 +29,7 @@ import {
   makeTransitionId,
 } from '@/app/_utils/lsMakers'
 import { RenderAddButton } from '../../_components/Svgs'
+import { RenderHearts } from './RenderHearts'
 
 // ------------------------Local Types ---------------------------------
 //input types for react-hook-form
@@ -37,15 +37,15 @@ type Inputs = {
   [key: `${number}`]: string //displayName
 }
 
-type MovementType = 'static' | 'transition'
-type MovementKeys = 'positions' | 'transitions'
+export type MovementType = 'static' | 'transition'
+export type MovementKeys = 'positions' | 'transitions'
 //-------------------------------Local Utils---------------------------------
 /**
  * looks in positions and transitions in Move to update the right value with slowRating
  * @param move Move
  * @returns Move
  */
-const getUpdatedMove = (
+export const getUpdatedMove = (
   move: Move,
   currentlyEditing: MovementType,
   movementGroup: MovementGroup,
@@ -146,161 +146,6 @@ const getPositionAndTransition = (
   }
 }
 
-/**
- *
- * Gets text for tooltip
- * @returns
- */
-const getText = (type: MovementType): string => {
-  switch (type) {
-    case 'static':
-      return 'For Footwork: statics are dancing using mainly the one position. Get confidence in the movement. Think big hip rotations. Shifting weight. Using hands and feet. Using knees. Adding flow concepts. Spinning. Shoulder alignment.'
-    case 'transition':
-      return 'For FW: Go from previous pose to current pose. Try whips with legs. Ease-in or Ease-out. Play with speed. Add foot steps inbetween. Alternate heel to toe. Butt slide. Foot slide. Hand slide.'
-  }
-}
-
-//-----------------------Renders ------------------------------
-
-/**
- * Renders 10 hearts. Used above each movement. Occurs multiple times per movement group.
- * @returns jsx
- */
-const RenderHearts = ({
-  rating,
-  move,
-  accessToLocalStorage,
-  currentlyEditing,
-  movementGroup,
-  setMove,
-  movements,
-}: {
-  movementGroup: MovementGroup
-  movements: MovementGroup[]
-  rating: number
-  move: Move
-  accessToLocalStorage: boolean
-  currentlyEditing: MovementType
-  setMove: Dispatch<SetStateAction<Move | null>>
-}) => {
-  //zustand to get state here rather than passed as props?
-
-  //make a movekey for when we update in localstorage. defaulting to positions
-  let moveKey: MovementKeys
-  switch (currentlyEditing) {
-    case 'static':
-      moveKey = 'positions'
-      break
-    case 'transition':
-      moveKey = 'transitions'
-      break
-  }
-
-  //----------------------------------------render-------------------------------
-  return (
-    <div className="flex flex-row-reverse items-center justify-end">
-      {
-        //render 10 hearts
-        Array.from(Array(10)).map((a, i) => {
-          return (
-            <>
-              <input
-                //When heart is clicked, the input will update local state and localstorage
-                onChange={(e) => {
-                  //-----------------------makes the updated move------------------
-                  console.log('making updated move')
-                  const updatedMove = getUpdatedMove(
-                    move,
-                    currentlyEditing,
-                    movementGroup,
-                    Number(e.target.id),
-                    movements,
-                  )
-
-                  console.log('updatedMove: ', updatedMove)
-                  //-----------------------updates display----------------------------
-
-                  //updates view, otherwise user has to refresh to get updates from localstorageDB data
-                  // setLocalMovements(updatedMove)
-                  setMove(updatedMove)
-
-                  //------------------------updates db--------------------------------
-
-                  //expression for if Move[] matches has a match provided moveid
-                  const matchCriteria = (a: Move) => a.moveId === move.moveId
-
-                  //all the moves from localstorage
-                  const globalMoves =
-                    getLocalStorageGlobal[lsUserLearning](accessToLocalStorage)
-
-                  //validation if local moveId exists in global moveId
-                  if (globalMoves.find(matchCriteria)) {
-                    //updates localstorage on click
-                    setLocalStorageGlobal[lsUserLearning](
-                      globalMoves.map((ogMove: Move) =>
-                        matchCriteria(ogMove) ? updatedMove : ogMove,
-                      ),
-                      accessToLocalStorage,
-                    )
-                  } else {
-                    //TODO have UI visible error handling
-                    console.log('cannot find moveid in localstorage')
-                  }
-                }}
-                checked={i === 10 - rating}
-                type="radio"
-                className="peer -ms-5 size-5 cursor-pointer
-              appearance-none border-0 bg-transparent
-              text-transparent
-              checked:bg-none focus:bg-none focus:ring-0 focus:ring-offset-0"
-                id={'' + (10 - i)}
-              />
-              <label
-                className="pointer-events-none text-gray-300 
-            peer-checked:text-red-500"
-              >
-                <svg
-                  className="size-5 flex-shrink-0"
-                  width="16"
-                  height="16"
-                  fill="currentColor"
-                  viewBox="0 0 16 16"
-                >
-                  <path
-                    fill-rule="evenodd"
-                    d="M8 1.314C12.438-3.248 23.534 4.735 8 15-7.534
-                 4.736 3.562-3.248 8 1.314z"
-                  ></path>
-                </svg>
-              </label>
-            </>
-          )
-        })
-      }
-    </div>
-  )
-}
-
-/** Renders Tooltips depending on what is being hovered */
-const RenderTooltip = ({ type }: { type: MovementType }) => {
-  return <></>
-  //FEATURE Have tooltips explain what statics and transitions are
-
-  //text of the tooltip
-  const text = getText(type)
-
-  //tooltip has been selected
-  const [isSelected, setIsSelected] = useState<boolean>(false)
-
-  //render
-  return (
-    <div className="flex items-center pl-0.5">
-      <svg width="15" height="15" viewBox="0 0 24 24">
-        <path d="m13 17-2 0 0-6 2 0 0 6zm-1-15c6 0 10 4 10 10s-4 10-10 10-10-4-10-10 4-10 10-10zm0 18c4 0 8-4 8-8s-4-8-8-8-8 4-8 8 4 8 8 8zm1-11-2 0 0-2 2 0 0 2z" />
-      </svg>
-    </div>
-  )
-}
 
 /**
  * Renders the heading text, and all the moves
@@ -315,6 +160,7 @@ const RenderMoveLearn = () => {
   const [localMovements, setLocalMovements] = useState<MovementGroup[]>([])
   const [accessToLocalStorage, setAccessToLocalStorage] = useState(false)
   const [move, setMove] = useState<Move | null>(null)
+  const [hasOppositeSide, setHasOppositeSide] = useState<boolean>(false)
 
   const searchParams = useSearchParams()
   const moveId: string | null = searchParams?.get('moveId') || null
@@ -327,7 +173,7 @@ const RenderMoveLearn = () => {
   useLocalStorage(setAccessToLocalStorage)
 
   //Hook to update after localstorage has been set
-  useEffect(() => {}, [setIsEditing])
+  useEffect(() => { }, [setIsEditing])
 
   //sets the order of the movements
   useEffect(() => {
@@ -539,6 +385,7 @@ const RenderMoveLearn = () => {
           <div className="mx-auto text-base text-xs leading-relaxed lg:w-2/3"></div>
         </div>
         <div>
+          {/* RenderMovementsGroup */}
           {move &&
             localMovements &&
             localMovements.map((movement, i) => {
@@ -547,134 +394,74 @@ const RenderMoveLearn = () => {
                 //-----makes a defaults if none found to handle edge cases----
                 //do not have a default for the last movementgroup as it's just a transition loop to repeat and doesnt have positions
                 position = i !== localMovements.length - 1 &&
-                  makeDefaultPosition({
-                    displayName: 'new-position',
-                  }),
+                makeDefaultPosition({
+                  displayName: 'new-position',
+                }),
                 //doesn't make a transitionobj for the first pos, as nothing to transition from
                 transition = i !== 0 &&
-                  makeDefaultTransition({
-                    displayName: 'new-transition',
-                    from: localMovements[i - 1].positionId || makePositionId(),
-                    to: position ? position.positionId : makePositionId(),
-                    transitionId: makeTransitionId(),
-                  }),
+                makeDefaultTransition({
+                  displayName: 'new-transition',
+                  from: localMovements[i - 1].positionId || makePositionId(),
+                  to: position ? position.positionId : makePositionId(),
+                  transitionId: makeTransitionId(),
+                }),
               } = getPositionAndTransition(movement, move)
               //---------------------------------------------------------------
-              return (
-                <div
-                  className="my-6 flex flex-col items-center"
-                  key={movement.displayName}
-                >
-                  <div className="flex">
-                    {
-                      //--------------MOVEMENT GROUP TITLE-------
-                      //if user is not editing, show delete and add button
-                      (isEditing !== null && isEditing[i]) || (
-                        <>
-                          <h1 className="title-font text-lg font-medium capitalize text-gray-900 dark:text-white">
-                            {movement.displayName}
-                          </h1>
-                          {/* ----------MODIFICATION BUTTONS-----*/}
-                          <div className="ml-2 flex items-center">
-                            <div className="w-2">
-                              <RenderEditButton
-                                onClick={() => {
-                                  setIsEditing({ [i]: true })
-                                }}
-                              />
-                            </div>
-                            <div className="ml-2 w-2">
-                              <RenderAddButton
-                                id={movement.movementId}
-                                onClick={onClickAddMovement}
-                              />
-                            </div>
-                            {
-                              //if there's more than one mvmt left, show delete button
-                              localMovements.length > 1 && (
-                                <div className="ml-2 w-2">
-                                  <RenderRedDeleteButton
-                                    id={movement.movementId}
-                                    onClick={onClickDeleteMovement}
-                                  />
-                                </div>
-                              )
-                            }
-                          </div>
-                        </>
-                      )
-                    }
-                    {
-                      // ---------------EDIT INPUT------------
-                      //if user is editing, edit button can save. dont show delete and add button.
-                      isEditing !== null && isEditing[i] && (
-                        <>
-                          <form onSubmit={handleSubmit(onSubmitNewMoveName(i))}>
-                            <DefaultStyledInput
-                              registerName={`${i}`}
-                              defaultValue={movement.displayName}
-                              register={register}
-                            />
-                            <button type="submit">
-                              <div className="ml-1 w-2">
-                                <RenderEditButton />
-                              </div>
-                            </button>
-                          </form>
-                        </>
-                      )
-                    }
-                  </div>
-                  <div className="mb-4 mt-2 flex justify-center">
-                    <div className="inline-flex h-1 w-16 rounded-full bg-indigo-500"></div>
-                  </div>
-                  {/*---------------TRANSITIONS w/ HEARTS & POSITIONS w/ HEARTS----- */}
-                  <div className="flex flex-col items-center text-xs">
-                    {transition && (
-                      //If its a transition render Transition
-                      <div className="flex flex-col py-3">
-                        <span>
-                          <RenderHearts
-                            movements={localMovements}
-                            setMove={setMove}
-                            accessToLocalStorage={accessToLocalStorage}
-                            rating={transition?.slowRating}
-                            move={move}
-                            currentlyEditing={'transition'}
-                            movementGroup={movement}
-                          />
-                          {'Transition: '}
-                          {transition.displayName}
-                        </span>
-                        <div>{'Practice: Slow Transitions'}</div>
-                      </div>
-                    )}
-                    {position && (
-                      //If its a position render position
-                      <div className="flex flex-col py-3">
-                        <span>
-                          <RenderHearts
-                            movements={localMovements}
-                            setMove={setMove}
-                            accessToLocalStorage={accessToLocalStorage}
-                            rating={position?.slowRating}
-                            move={move}
-                            currentlyEditing={'static'}
-                            movementGroup={movement}
-                          />
-                          {'Position: '}
-                          {position.displayName}
-                        </span>
-                        <span className="flex">
-                          <div>{'Practice: Slow Statics'}</div>
-                          <RenderTooltip type={'static'} />
-                        </span>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              )
+              // TODO refactor this to use zustand. looks ugly like this and isn't clear in rendermovementgroup what all the functions do
+              return <RenderMovementGroup movement={movement} key={movement.displayName} position={position} isEditing={isEditing} onClickAddMovement={onClickAddMovement} transition={transition} setIsEditing={setIsEditing} indexNumber={i} localMovements={localMovements} handleSubmit={handleSubmit} onSubmitNewMoveName={onSubmitNewMoveName} register={register} setMove={setMove} accessToLocalStorage={accessToLocalStorage} move={move}
+              />
             })}
+        </div>
+        <form className="py-10">
+          <span>
+            <input type="checkbox" />
+            Move into reverse direction?
+          </span>
+          <div className="text-xs">
+            this will double the movement groups
+          </div>
+        </form>
+        <div>
+          {move &&
+            localMovements && hasOppositeSide &&
+            localMovements.map((movement, i) => {
+              //gets position and transition referred to by movementGroup obj
+              const {
+                //-----makes a defaults if none found to handle edge cases----
+                //do not have a default for the last movementgroup as it's just a transition loop to repeat and doesnt have positions
+                position = i !== localMovements.length - 1 &&
+                makeDefaultPosition({
+                  displayName: 'new-position',
+                }),
+                //doesn't make a transitionobj for the first pos, as nothing to transition from
+                transition = i !== 0 &&
+                makeDefaultTransition({
+                  displayName: 'new-transition',
+                  from: localMovements[i - 1].positionId || makePositionId(),
+                  to: position ? position.positionId : makePositionId(),
+                  transitionId: makeTransitionId(),
+                }),
+              } = getPositionAndTransition(movement, move)
+              //---------------------------------------------------------------
+              return <RenderMovementGroup
+                movement={movement}
+                key={movement.displayName}
+                position={position}
+                isEditing={isEditing}
+                onClickAddMovement={onClickAddMovement}
+                transition={transition}
+                setIsEditing={setIsEditing}
+                indexNumber={i}
+                localMovements={localMovements}
+                handleSubmit={handleSubmit}
+                onSubmitNewMoveName={onSubmitNewMoveName}
+                register={register}
+                setMove={setMove}
+                accessToLocalStorage={accessToLocalStorage}
+                move={move}
+              />
+            })}
+
         </div>
       </div>
     </section>
